- Implementata chiamata HTTP di tipo GET al servizio API Uniparthenope/v1/students/totalExams per ottenere informazioni sul totale degli esami superati e del totale dei CFU dello studente. Il parametro {matId} dello studente è stato passato staticamente per scopi di testing in attesa dell'aggiornamento del modello User. - Inoltre, sono state apportate modifiche a diversi widget per aderire al pattern userRoles in modo da semplificarne l'utilizzo futuro.

+++ appuniparthenope/lib/model/student_carrer_data.dart
@@ -3,8 +3,8 @@ import 'dart:convert';
 class TotalExamStudent {
   int totAdSuperate;
   int numAdSuperate;
-  int cfuPar;
-  int cfuTot;
+  double cfuPar;
+  double cfuTot;
 
   TotalExamStudent({
     required this.totAdSuperate,
